third_party/blink/web_tests/http/tests/worklet/animation-worklet-csp-eval-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of undefined (reading 'addModule')
